First draft of the const package
